new: delete endpoint

+++ src/main/java/com/codacy/demo/EmployeeRepository.java
@@ -6,8 +6,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EmployeeRepository extends CrudRepository<Employee, Integer> {
-
+public interface EmployeeRepository extends CrudRepository<Employee, Long> {
     @Override
     List<Employee> findAll();
 }
